# todo-list/api/api.yaml
swagger: '2.0'
info:
  version: '1.0.0'
  title: 'Todo App API'
paths:
  /tasks:
    get:
      summary: Get all tasks
      responses:
        '200':
          description: A list of tasks
          schema:
            type: array
            items:
              $ref: '#/definitions/TodoItem'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'

    post:
      summary: Create a new task
      parameters:
        - name: task
          in: body
          description: The task to create
          required: true
          schema:
            $ref: '#/definitions/TodoItem'
      responses:
        '201':
          description: Successfully created a new task
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'

  /tasks/{id}:
    get:
      summary: Retrieve a task
      parameters:
        - name: id
          in: path
          description: Task ID
          required: true
          type: string
      responses:
        '200':
          description: A task
          schema:
            $ref: '#/definitions/TodoItem'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'

    put:
      summary: Update a task
      parameters:
        - name: id
          in: path
          description: Task ID
          required: true
          type: string
        - name: task
          in: body
          description: Task to update
          required: true
          schema:
            $ref: '#/definitions/TodoItem'
      responses:
        '200':
          description: Successfully updated the task
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'

    delete:
      summary: Delete a task
      parameters:
        - name: id
          in: path
          description: Task ID to delete
          required: true
          type: string
      responses:
        '200':
          description: Successfully deleted the task
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'

definitions:
  TodoItem:
    required:
      - id
      - task
      - is_completed
    properties:
      id:
        type: string
      task:
        type: string
      is_completed:
        type: boolean

  Error:
    required:
      - message
    properties:
      message:
        type: string